+++ tests/admin/CoreAdminTest.php
@@ -1,4 +1,5 @@
 <?php
+
 use PHPUnit\Framework\TestCase;
 require './vendor/autoload.php';
 require_once 'Modules/core/Controller/CoreconnectionController.php';
@@ -16,76 +17,51 @@ require_once 'Modules/core/Model/CoreMainSubMenu.php';
 require_once 'Modules/core/Model/CoreMainMenuItem.php';
 require_once 'Modules/core/Model/CorePendingAccount.php';
 
+require_once 'tests/BaseTest.php';
 
-class CoreTest extends TestCase
-{
-    private static $allSpaces = [];
-    private static $allUsers = [];
 
-    private function asAdmin($id_space=0) {
-        return $this->asUser(Configuration::get('admin_user', 'pfmadmin'), $id_space);
-    }
-
-    private function asUser($name, $id_space) {
-        for($i=0;$i<count(self::$allUsers);$i++) {
-            if(self::$allUsers[$i]['name'] == $name) {
-                $_SESSION['id_user'] = self::$allUsers[$i]['id'];
-                $_SESSION['id_space'] = $id_space;
-                $_SESSION["user_settings"] = ["language" => "en"];                
-                return self::$allUsers[$i];
-            }
-        }
-        return null;
-    }
+class CoreTest extends BaseTest {
 
     public function testInstallAndCoreAccess()
     {
         $req = new Request(["path" => ""], true);
         $c = new CoreconnectionController($req);
-        $res = $c->indexAction();
-        $this->assertTrue(isset($res['metadesc']));
-
+        $c->indexAction();
         $m = new CoreUser();
         $users = $m->getAll();
-        foreach($users as $user){
-            self::$allUsers[] = ["name" => $user['login'], "id" => intval($user['id'])];
-        }
-
-
+        Configuration::getLogger()->debug('[core] users', ['users' => $users]);
+        $this->assertTrue($users && !empty($users));
     }
 
     public function testCreateSpace() {
-        $spaceName = uniqid();
-        $this->asAdmin();
-        $req = new Request([
-            "path" => "spaceadminedit/0",
-            "formid" => "corespaceadminedit",
-            "name" => $spaceName,
-            "contact" => "admin",
-            "status" => 1,
-            "color" => "#154186",
-            "support" => "",
-            "description" => "",
-            "admins" => []
-        ], true);
-        $c = new CorespaceadminController($req);
-        $c->editAction(0);
-        $m = new CoreSpace();
-        $spaces = $m->getSpaces('id');
-        $spaceId = 0;
-        foreach($spaces as $space){
-            if ($space['name'] == $spaceName) {
-                $spaceId = $space['id'];
-            }
+        $ctx = $this->Context();
+        $spaces = $ctx['spaces'];
+        foreach($spaces as $spaceName => $data) {
+            $this->asAdmin();
+            $req = new Request([
+                "path" => "spaceadminedit/0",
+                "formid" => "corespaceadminedit",
+                "name" => $spaceName,
+                "contact" => "admin",
+                "status" => 1,
+                "color" => "#000000",
+                "txtcolor" => "#ffffff",
+                "support" => "",
+                "description" => "",
+                "admins" => []
+            ], true);
+            $c = new CorespaceadminController($req);
+            $data = $c->editAction(0);
+            $newSpace = $data['space'];
+            $this->assertTrue($newSpace['id'] > 0);
+            $this->assertNotFalse($this->space($spaceName));
         }
-        $this->assertTrue($spaceId > 0);
-        self::$allSpaces[] = ["name" => $spaceName, "id" => $spaceId];
     }
 
     public function testCreateMenus() {
         $this->asAdmin();
 
-        $menuName = uniqid();
+        $menuName = 'menu1';
         $req = new Request([
             "path" => "coremainmenuedit/0",
             "formid" => "editmainmenuform",
@@ -98,7 +74,7 @@ class CoreTest extends TestCase
         $menuModel = new CoreMainMenu();
         $menus = $menuModel->getAll();
 
-        $subMenuName = uniqid();
+        $subMenuName = 'submenu1';
         $req = new Request([
             "path" => "coresubmenuedit/0",
             "formid" => "editmainsubmenuform",
@@ -112,8 +88,9 @@ class CoreTest extends TestCase
         $subMenuModel = new CoreMainSubMenu();
         $subMenus = $subMenuModel->getAll();
 
-        foreach (self::$allSpaces as $space) {
-            $itemMenuName = uniqid();
+        $spaces = $this->spaces();
+        foreach ($spaces as $space) {
+            $itemMenuName = $space['name'];
             $req = new Request([
                 "path" => "coremainmenuitemedit/0",
                 "formid" => "editmenuitemform",
@@ -131,61 +108,108 @@ class CoreTest extends TestCase
     }
 
     public function testCreateUser() {
-        $space = self::$allSpaces[0];
-        $userName = uniqid();
         $this->asAdmin();
-        $req = new Request([
-            "path" => "corespaceaccessuseradd/".$space['id'],
-            "formid" => "createuseraccountform",
-            "name" => $userName,
-            "firstname" => $userName,
-            "login" =>  $userName,
-            "email" => "",
-            "phone" => ""
-        ], true);
-        $c = new CorespaceaccessController($req);
-        $c->useraddAction($space['id']);
-        $userId = 0;
-        $m = new CoreUser();
-        $users = $m->getAll();
-        $userId = 0;
-        foreach($users as $user){
-            if ($user['name'] == $userName) {
-                $userId = $user['id'];
+        $ctx = $this->Context();
+        $spaces = $ctx['spaces'];
+
+        foreach($spaces as $spaceName => $spaceMembers) {
+            $usersToAdd = [];
+            $space = $this->space($spaceName);
+            foreach($spaceMembers['admins'] as $u) {
+                $usersToAdd[] = $u;
+            }
+            foreach($spaceMembers['managers'] as $u) {
+                $usersToAdd[] = $u;
+            }
+            foreach($spaceMembers['users'] as $u) {
+                $usersToAdd[] = $u;
+            }
+            foreach($usersToAdd as $userName) {
+                $req = new Request([
+                    "path" => "corespaceaccessuseradd/".$space['id'],
+                    "formid" => "createuseraccountform",
+                    "name" => $userName,
+                    "firstname" => $userName,
+                    "login" =>  $userName,
+                    "email" => $userName."@pfm.org",
+                    "phone" => ""
+                ], true);
+                $c = new CorespaceaccessController($req);
+                $data = $c->useraddAction($space['id']);
+                $newUser = $data['user'];
+                $this->assertTrue($newUser && $newUser['id'] > 0);
+    
+                $sm = new CoreSpace();
+                $role = $sm->getUserSpaceRole($space['id'], $newUser['id']);
+                $this->assertEquals(-1, $role);
+                $pm = new CorePendingAccount();
+                $this->assertTrue($pm->isActuallyPending($space['id'], $newUser['id']));
+
             }
         }
-        $this->assertTrue($userId > 0);
-        self::$allUsers[] = ["name" => $userName, "id" => intval($userId)];
+    }
 
-        $sm = new CoreSpace();
-        $role = $sm->getUserSpaceRole($space['id'], $userId);
-        $this->assertEquals(-1, $role);
+
+    public function activatePendingAs($space_id, $user_id, $role) {
         $pm = new CorePendingAccount();
-        $this->assertTrue($pm->isActuallyPending($space['id'], $userId));
+        $pendings = $pm->getBySpaceIdAndUserId($space_id, $user_id);
+
+        $req = new Request([
+            "path" => "corespacependinguseredit/".$space_id."/".$pendings['id'],
+            "formid" => "pendingusereditactionform",
+            "role" => $role
+        ], true);
+
+        $c = new CorespaceaccessController($req);
+        $c->pendingusereditAction($space_id, $pendings['id']);
+        $sm = new CoreSpace();
+        $urole = $sm->getUserSpaceRole($space_id, $user_id);
+        $this->assertEquals($role, $urole);
+        $pm = new CorePendingAccount();
+        $this->assertFalse($pm->isActuallyPending($space_id, $user_id));
     }
 
     public function testActivatePendingUserAsUser() {
         $this->asAdmin();
+        $ctx = $this->Context();
+        $spaces = $ctx['spaces'];
 
-        $space = self::$allSpaces[0];
-        $user = self::$allUsers[1];
-        $pm = new CorePendingAccount();
-        $pendings = $pm->getBySpaceIdAndUserId($space['id'], $user['id']);
-
-        $req = new Request([
-            "path" => "corespacependinguseredit/".$space['id']."/".$pendings['id'],
-            "formid" => "pendingusereditactionform",
-            "role" => CoreSpace::$USER
-        ], true);
-
-
-        $c = new CorespaceaccessController($req);
-        $c->pendingusereditAction($space['id'], $pendings['id']);
-        $sm = new CoreSpace();
-        $role = $sm->getUserSpaceRole($space['id'], $user['id']);
-        $this->assertEquals(CoreSpace::$USER, $role);
-        $pm = new CorePendingAccount();
-        $this->assertFalse($pm->isActuallyPending($space['id'], $user['id']));
+        foreach($spaces as $spaceName => $spaceMembers) {
+            $space = $this->space($spaceName);
+            foreach($spaceMembers['users'] as $u) {
+               $user = $this->user($u);
+               $this->activatePendingAs($space['id'], $user['id'], CoreSpace::$USER);
+            }
+        }
     }
+
+    public function testActivatePendingUserAsSpaceManager() {
+        $this->asAdmin();
+        $ctx = $this->Context();
+        $spaces = $ctx['spaces'];
+
+        foreach($spaces as $spaceName => $spaceMembers) {
+            $space = $this->space($spaceName);
+            foreach($spaceMembers['managers'] as $u) {
+               $user = $this->user($u);
+               $this->activatePendingAs($space['id'], $user['id'], CoreSpace::$MANAGER);
+            }
+        }
+    }
+
+    public function testActivatePendingUserAsSpaceAdmin() {
+        $this->asAdmin();
+        $ctx = $this->Context();
+        $spaces = $ctx['spaces'];
+
+        foreach($spaces as $spaceName => $spaceMembers) {
+            $space = $this->space($spaceName);
+            foreach($spaceMembers['admins'] as $u) {
+               $user = $this->user($u);
+               $this->activatePendingAs($space['id'], $user['id'], CoreSpace::$ADMIN);
+            }
+        }
+    }
+
 }